Deduplicate tx tracer
