+++ packages/keplr-cosmos/src/tx-tracer/index.ts
@@ -1,13 +1,17 @@
-import { TxEventMap, WsReadyState } from "./types";
+import { Buffer } from "buffer";
 
-import { Buffer } from "buffer/";
+import { TxEventMap, WsReadyState } from "./types";
 
 type Listeners = {
   [K in keyof TxEventMap]?: TxEventMap[K][];
 };
 
+/**
+ * TendermintTxTracer is almost same with the `TendermintTxTracer` in the @osmosis-labs/keplr-cosmos library.
+ * Changes for some mistake on the original `TendermintTxTracer`.
+ */
 export class TendermintTxTracer {
-  protected ws: WebSocket;
+  protected ws!: WebSocket;
 
   protected newBlockSubscribes: {
     handler: (block: any) => void;
@@ -16,7 +20,7 @@ export class TendermintTxTracer {
   protected txSubscribes: Map<
     number,
     {
-      hash: Uint8Array;
+      params: Record<string, string | number | boolean>;
       resolver: (data?: unknown) => void;
       rejector: (e: Error) => void;
     }
@@ -27,7 +31,7 @@ export class TendermintTxTracer {
     number,
     {
       method: string;
-      params: unknown[];
+      params: Record<string, string | number | boolean>;
       resolver: (data?: unknown) => void;
       rejector: (e: Error) => void;
     }
@@ -42,12 +46,7 @@ export class TendermintTxTracer {
       wsObject?: new (url: string, protocols?: string | string[]) => WebSocket;
     } = {}
   ) {
-    this.ws = this.options.wsObject
-      ? new this.options.wsObject(this.getWsEndpoint())
-      : new WebSocket(this.getWsEndpoint());
-    this.ws.onopen = this.onOpen;
-    this.ws.onmessage = this.onMessage;
-    this.ws.onclose = this.onClose;
+    this.open();
   }
 
   protected getWsEndpoint(): string {
@@ -66,8 +65,25 @@ export class TendermintTxTracer {
     return url;
   }
 
+  open() {
+    this.ws = this.options.wsObject
+      ? new this.options.wsObject(this.getWsEndpoint())
+      : new WebSocket(this.getWsEndpoint());
+    this.ws.onopen = this.onOpen;
+    this.ws.onmessage = this.onMessage;
+    this.ws.onclose = this.onClose;
+  }
+
   close() {
     this.ws.close();
+  }
+
+  get numberOfSubscriberOrPendingQuery(): number {
+    return (
+      this.newBlockSubscribes.length +
+      this.txSubscribes.size +
+      this.pendingQueries.size
+    );
   }
 
   get readyState(): WsReadyState {
@@ -105,7 +121,7 @@ export class TendermintTxTracer {
     }
 
     for (const [id, tx] of this.txSubscribes) {
-      this.sendSubscribeTxRpc(id, tx.hash);
+      this.sendSubscribeTxRpc(id, tx.params);
     }
 
     for (const [id, query] of this.pendingQueries) {
@@ -134,14 +150,8 @@ export class TendermintTxTracer {
                 .get(obj.id)!
                 .rejector(new Error(obj.error.data || obj.error.message));
             } else {
-              // XXX: I'm not sure why this happens, but somtimes the form of tx id delivered under the "tx_result" field.
-              if (obj.result?.tx_result) {
-                // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-                this.pendingQueries.get(obj.id)!.resolver(obj.result.tx_result);
-              } else {
-                // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-                this.pendingQueries.get(obj.id)!.resolver(obj.result);
-              }
+              // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+              this.pendingQueries.get(obj.id)!.resolver(obj.result);
             }
 
             this.pendingQueries.delete(obj.id);
@@ -174,7 +184,7 @@ export class TendermintTxTracer {
           }
         }
       } catch (e: any) {
-        console.log(
+        console.error(
           `Tendermint websocket jsonrpc response is not JSON: ${
             e.message || e.toString()
           }`
@@ -189,7 +199,13 @@ export class TendermintTxTracer {
     }
   };
 
-  subscribeBlock(handler: (block: any) => void) {
+  /**
+   * SubscribeBlock receives the handler for the block.
+   * The handelrs shares the subscription of block.
+   * @param handler
+   * @return unsubscriber
+   */
+  subscribeBlock(handler: (block: any) => void): () => void {
     this.newBlockSubscribes.push({
       handler,
     });
@@ -197,6 +213,12 @@ export class TendermintTxTracer {
     if (this.newBlockSubscribes.length === 1) {
       this.sendSubscribeBlockRpc();
     }
+
+    return () => {
+      this.newBlockSubscribes = this.newBlockSubscribes.filter(
+        (s) => s.handler !== handler
+      );
+    };
   }
 
   protected sendSubscribeBlockRpc(): void {
@@ -205,7 +227,7 @@ export class TendermintTxTracer {
         JSON.stringify({
           jsonrpc: "2.0",
           method: "subscribe",
-          params: ["tm.event='NewBlock'"],
+          params: { query: "tm.event='NewBlock'" },
           id: 1,
         })
       );
@@ -213,64 +235,141 @@ export class TendermintTxTracer {
   }
 
   // Query the tx and subscribe the tx.
-  traceTx(hash: Uint8Array): Promise<any> {
+  traceTx(
+    query: Uint8Array | Record<string, string | number | boolean>
+  ): Promise<any> {
     return new Promise<any>((resolve) => {
       // At first, try to query the tx at the same time of subscribing the tx.
       // But, the querying's error will be ignored.
-      this.queryTx(hash)
-        .then(resolve)
+      this.queryTx(query)
+        .then((result) => {
+          if (query instanceof Uint8Array) {
+            resolve(result);
+            return;
+          }
+
+          if (result?.total_count !== "0") {
+            resolve(result);
+            return;
+          }
+        })
         .catch(() => {
           // noop
         });
 
-      this.subscribeTx(hash).then(resolve);
-    }).then((tx) => {
-      // Occasionally, even if the subscribe tx event occurs, the state through query is not changed yet.
-      // Perhaps it is because the block has not been committed yet even though the result of deliverTx in tendermint is complete.
-      // This method is usually used to reflect the state change through query when tx is completed.
-      // The simplest solution is to just add a little delay.
-      return new Promise((resolve) => {
-        setTimeout(() => resolve(tx), 100);
-      });
+      this.subscribeTx(query).then(resolve);
     });
   }
 
-  subscribeTx(hash: Uint8Array): Promise<any> {
-    const id = this.createRandomId();
+  subscribeTx(
+    query: Uint8Array | Record<string, string | number | boolean>
+  ): Promise<any> {
+    if (query instanceof Uint8Array) {
+      const id = this.createRandomId();
 
-    return new Promise<unknown>((resolve, reject) => {
-      this.txSubscribes.set(id, {
-        hash,
-        resolver: resolve,
-        rejector: reject,
+      const params = {
+        query: `tm.event='Tx' AND tx.hash='${Buffer.from(query)
+          .toString("hex")
+          .toUpperCase()}'`,
+      };
+
+      return new Promise<unknown>((resolve, reject) => {
+        this.txSubscribes.set(id, {
+          params,
+          resolver: resolve,
+          rejector: reject,
+        });
+
+        this.sendSubscribeTxRpc(id, params);
       });
+    } else {
+      const id = this.createRandomId();
 
-      this.sendSubscribeTxRpc(id, hash);
-    });
+      const params = {
+        query:
+          `tm.event='Tx' and ` +
+          Object.keys(query)
+            .map((key) => {
+              return {
+                key,
+                value: query[key],
+              };
+            })
+            .map((obj) => {
+              return `${obj.key}=${
+                typeof obj.value === "string" ? `'${obj.value}'` : obj.value
+              }`;
+            })
+            .join(" and "),
+        page: "1",
+        per_page: "1",
+        order_by: "desc",
+      };
+
+      return new Promise<unknown>((resolve, reject) => {
+        this.txSubscribes.set(id, {
+          params,
+          resolver: resolve,
+          rejector: reject,
+        });
+
+        this.sendSubscribeTxRpc(id, params);
+      });
+    }
   }
 
-  protected sendSubscribeTxRpc(id: number, hash: Uint8Array): void {
+  protected sendSubscribeTxRpc(
+    id: number,
+    params: Record<string, string | number | boolean>
+  ): void {
     if (this.readyState === WsReadyState.OPEN) {
       this.ws.send(
         JSON.stringify({
           jsonrpc: "2.0",
           method: "subscribe",
-          params: [
-            `tm.event='Tx' AND tx.hash='${Buffer.from(hash)
-              .toString("hex")
-              .toUpperCase()}'`,
-          ],
+          params: params,
           id,
         })
       );
     }
   }
 
-  queryTx(hash: Uint8Array): Promise<any> {
-    return this.query("tx", [Buffer.from(hash).toString("base64"), false]);
+  queryTx(
+    query: Uint8Array | Record<string, string | number | boolean>
+  ): Promise<any> {
+    if (query instanceof Uint8Array) {
+      return this.query("tx", {
+        hash: Buffer.from(query).toString("base64"),
+        prove: false,
+      });
+    } else {
+      const params = {
+        query: Object.keys(query)
+          .map((key) => {
+            return {
+              key,
+              value: query[key],
+            };
+          })
+          .map((obj) => {
+            return `${obj.key}=${
+              typeof obj.value === "string" ? `'${obj.value}'` : obj.value
+            }`;
+          })
+          .join(" and "),
+        page: "1",
+        per_page: "1",
+        order_by: "desc",
+      };
+
+      return this.query("tx_search", params);
+    }
   }
 
-  protected query(method: string, params: unknown[]): Promise<any> {
+  protected query(
+    method: string,
+    params: Record<string, string | number | boolean>
+  ): Promise<any> {
     const id = this.createRandomId();
 
     return new Promise<unknown>((resolve, reject) => {
@@ -285,7 +384,11 @@ export class TendermintTxTracer {
     });
   }
 
-  protected sendQueryRpc(id: number, method: string, params: unknown[]) {
+  protected sendQueryRpc(
+    id: number,
+    method: string,
+    params: Record<string, string | number | boolean>
+  ) {
     if (this.readyState === WsReadyState.OPEN) {
       this.ws.send(
         JSON.stringify({
