+++ packages/keplr-hooks/package.json
@@ -26,8 +26,6 @@
     "@keplr-wallet/background": "0.10.24-ibc.go.v7.hot.fix",
     "@keplr-wallet/common": "0.10.24-ibc.go.v7.hot.fix",
     "@osmosis-labs/keplr-cosmos": "0.10.24-ibc.go.v7.hot.fix",
-    "@keplr-wallet/ens": "0.10.24-ibc.go.v7.hot.fix",
-    "@keplr-wallet/popup": "0.10.24-ibc.go.v7.hot.fix",
     "@keplr-wallet/proto-types": "0.10.24-ibc.go.v7.hot.fix",
     "@keplr-wallet/types": "0.10.24-ibc.go.v7.hot.fix",
     "@keplr-wallet/unit": "0.10.24-ibc.go.v7.hot.fix",
