+++ packages/stores/src/tx/index.ts
@@ -1,5 +1,3 @@
 export * from "./ibc";
 export * from "./prettify";
-export * from "./tracer";
-export * from "./types";
 export * from "./utils";
